+++ lib/improver/tests/nbhood/percentiles/test_CircularKernelNumpy.py
@@ -38,7 +38,7 @@ from iris.cube import Cube
 from iris.tests import IrisTest
 import numpy as np
 
-from improver.nbhood_percentiles import CircularKernelNumpy
+from improver.nbhood.percentiles import CircularKernelNumpy
 from improver.tests.nbhood.nbhood.test_NeighbourhoodProcessing import (
     set_up_cube, set_up_cube_lat_long)
 from improver.percentile import PercentileConverter
