+++ lib/improver/tests/nbhood/percentiles/test_NeighbourhoodPercentiles.py
@@ -39,7 +39,7 @@ from iris.tests import IrisTest
 import numpy as np
 
 
-from improver.nbhood_percentiles import NeighbourhoodPercentiles as NBHood
+from improver.nbhood.percentiles import NeighbourhoodPercentiles as NBHood
 from improver.tests.ensemble_calibration.ensemble_calibration.helper_functions\
     import add_forecast_reference_time_and_forecast_period
 from improver.tests.nbhood.nbhood.test_NeighbourhoodProcessing import (
